+++ src/storage/Cargo.toml
@@ -38,6 +38,7 @@ mz-persist-types = { path = "../persist-types", optional = true }
 mz-pgcopy = { path = "../pgcopy", optional = true }
 mz-postgres-util = { path = "../postgres-util", optional = true }
 mz-repr = { path = "../repr" }
+mz-service = { path = "../service" }
 mz-timely-util = { path = "../timely-util" }
 once_cell = { version = "1.12.0", optional = true }
 postgres-protocol = { git = "https://github.com/MaterializeInc/rust-postgres", optional = true }
