+++ src/storaged/src/bin/storaged.rs
@@ -17,6 +17,7 @@ use once_cell::sync::Lazy;
 use tracing::info;
 
 use mz_build_info::{build_info, BuildInfo};
+use mz_dataflow_types::client::proto_storage_server::ProtoStorageServer;
 use mz_dataflow_types::client::StorageClient;
 use mz_dataflow_types::connections::ConnectionContext;
 use mz_orchestrator_tracing::TracingCliArgs;
@@ -24,6 +25,7 @@ use mz_ore::cli::{self, CliConfig};
 use mz_ore::metrics::MetricsRegistry;
 use mz_ore::now::SYSTEM_TIME;
 use mz_pid_file::PidFile;
+use mz_service::grpc;
 
 // Disable jemalloc on macOS, as it is not well supported [0][1][2].
 // The issues present as runaway latency on load test workloads that are
@@ -170,11 +172,6 @@ async fn run(args: Args) -> Result<(), anyhow::Error> {
         ),
     };
 
-    let serve_config = mz_dataflow_types::client::grpc::ServeConfig {
-        listen_addr: args.listen_addr,
-        linger: args.linger,
-    };
-
     assert!(
         !args.reconcile,
         "Storage runtime does not support command reconciliation."
@@ -182,10 +179,13 @@ async fn run(args: Args) -> Result<(), anyhow::Error> {
     let (_server, client) = mz_storage::serve(config)?;
     let client: Box<dyn StorageClient> = Box::new(client);
 
-    mz_dataflow_types::client::grpc::serve(
-        serve_config,
+    mz_service::grpc::serve(
+        grpc::ServeConfig {
+            listen_addr: args.listen_addr,
+            linger: args.linger,
+        },
         client,
-        mz_dataflow_types::client::grpc::ProtoStorageServer::new,
+        ProtoStorageServer::new,
     )
     .await
 }
